backend/requirements.txt
langchain
spacy
PyMuPDF
langchain_community
langchain_ollama
langchain_core
langchain_huggingface
fastapi
pymongo
pydantic
langchain_groq
jose
passlib
librosa
numpy
certifi
python-multipart
whisper
openai-whisper
scipy
gtts
pydub
uvicorn[standard]
pydantic[email]
python-jose[cryptography]
sentence-transformers
argon2-cffi
